+++ cpmpy/expressions/python_builtins.py
@@ -19,11 +19,9 @@
         sum
 """
 import numpy as np
-from fishhook import *
-from .variables import NDVarArray, _IntVarImpl, _BoolVarImpl
+from .variables import NDVarArray
 from .core import Expression, Operator
 from .globalconstraints import Minimum, Maximum
-from ..exceptions import CPMpyException
 
 # Overwriting all/any python built-ins
 # all: listwise 'and'
@@ -102,39 +100,3 @@ def sum(iterable):
     if not any(isinstance(elem, Expression) for elem in iterable):
         return np.sum(iterable)
     return Operator("sum", iterable)
-
-@hook(bool, int)
-def __and__(self, other):
-    if isinstance(other, Expression):
-        return other.__rand__(self)
-    return orig(self, other)
-
-@hook(bool, int)
-def __or__(self, other):
-    if isinstance(other, Expression):
-        return other.__ror__(self)
-    return orig(self, other)
-
-@hook(bool, int)
-def implies(self, other):
-    if not (isinstance(other,Expression)):
-        return orig(self, other)
-    else:
-
-        if not self.is_bool():
-            raise CPMpyException(f"Logical implication involving an integer ({self}) is not allowed")
-
-        if isinstance(self,int):
-            self = (self == 1)  # convert to True or False
-
-        if not isinstance(other,_BoolVarImpl):
-            raise CPMpyException(f"Logical implication involving an IntVar ({other}) is not allowed")
-
-        return Operator("->", [self, other])
-
-@hook(int)
-def is_bool(self):
-    if self > 1 or self < 0:
-        return False
-    else:
-        return True
